# solutions/rsw-reference/RSW.openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "RSW Official",
    "version": "1.0.0",
    "contact": {}
  },
  "paths": {
...
    "/projects/{projectId}/sshKeys": {
      "get": {
        "tags": ["sshKeys"],
        "summary": "List User SSH Keys for Project",
        "description": "List User SSH Keys for Project",
        "operationId": "listUserSshKeysForProject",
        "parameters": [
          {
            "name": "Cookie",
            "in": "header",
            "schema": {
              "type": "string",
              "example": "access_token={{ACCESS_TOKEN}};_csrf={{CSRF_COOKIE}}"
            }
          },
          {
            "name": "csrf-token",
            "in": "header",
            "schema": {
              "type": "string",
              "example": "{{CSRF_TOKEN}}"
            }
          }
        ],
        "responses": {
          "200": {}
        }
      },
      "post": {
        "tags": ["sshKeys"],
        "summary": "Create SSH Key",
        "description": "Create SSH Key",
        "operationId": "createSshKey",
        "parameters": [
          {
            "name": "Cookie",
            "in": "header",
            "schema": {
              "type": "string",
              "example": "access_token={{ACCESS_TOKEN}};_csrf={{CSRF_COOKIE}}"
            }
          },
          {
            "name": "csrf-token",
            "in": "header",
            "schema": {
              "type": "string",
              "example": "{{CSRF_TOKEN}}"
            }
          }
        ],
        "responses": {
          "201": {}
        }
      },
      "parameters": [
        {
          "name": "projectId",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string",
            "example": ""
          }
        }
      ]
    },
    "/projects/{projectId}/sshKeys/{sshKeyId}/purge": {
      "delete": {
        "tags": ["sshKeys"],
        "summary": "Delete SSH Key",
        "description": "Delete SSH Key",
        "operationId": "deleteSshKey",
        "parameters": [
          {
            "name": "Cookie",
            "in": "header",
            "schema": {
              "type": "string",
              "example": "access_token={{ACCESS_TOKEN}};_csrf={{CSRF_COOKIE}}"
            }
          },
          {
            "name": "csrf-token",
            "in": "header",
            "schema": {
              "type": "string",
              "example": "{{CSRF_TOKEN}}"
            }
          }
        ],
        "responses": {
          "204": {}
        }
      },
      "parameters": [
        {
          "name": "projectId",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string",
            "example": ""
          }
        },
        {
          "name": "sshKeyId",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string",
            "example": ""
          }
        }
      ]
    },
    "/projects/{projectId}/environments/{environmentId}/sshKeys": {
      "get": {
        "tags": ["sshKeys"],
        "summary": "Send SSH Public Key",
        "description": "Send SSH Public Key",
        "operationId": "sendSshPublicKey",
        "parameters": [
          {
            "name": "Cookie",
            "in": "header",
            "schema": {
              "type": "string",
              "example": "access_token={{ACCESS_TOKEN}};_csrf={{CSRF_COOKIE}}"
            }
          },
          {
            "name": "csrf-token",
            "in": "header",
            "schema": {
              "type": "string",
              "example": "{{CSRF_TOKEN}}"
            }
          }
        ],
        "responses": {
          "200": {}
        }
      },
      "parameters": [
        {
          "name": "projectId",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string",
            "example": ""
          }
        },
        {
          "name": "environmentId",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string",
            "example": ""
          }
        }
      ]
    },
...
  },
...
}